+++ backend/packages/binary-generators/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @cat-factory/binary-generators
 
+## 0.3.47
+
+### Patch Changes
+
+- Updated dependencies [b75fa3c]
+  - @cat-factory/contracts@0.352.0
+  - @cat-factory/kernel@0.345.0
+
+## 0.3.46
+
+### Patch Changes
+
+- Updated dependencies [bba4beb]
+  - @cat-factory/kernel@0.344.0
+
 ## 0.3.45
 
 ### Patch Changes