+++ backend/packages/eks/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @cat-factory/eks
 
+## 0.1.386
+
+### Patch Changes
+
+- Updated dependencies [b75fa3c]
+  - @cat-factory/contracts@0.352.0
+  - @cat-factory/kernel@0.345.0
+  - @cat-factory/integrations@0.172.16
+
+## 0.1.385
+
+### Patch Changes
+
+- Updated dependencies [bba4beb]
+  - @cat-factory/kernel@0.344.0
+  - @cat-factory/integrations@0.172.15
+
 ## 0.1.384
 
 ### Patch Changes