+++ backend/packages/gitlab/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @cat-factory/gitlab
 
+## 0.23.10
+
+### Patch Changes
+
+- Updated dependencies [b75fa3c]
+  - @cat-factory/contracts@0.352.0
+  - @cat-factory/kernel@0.345.0
+
+## 0.23.9
+
+### Patch Changes
+
+- Updated dependencies [bba4beb]
+  - @cat-factory/kernel@0.344.0
+
 ## 0.23.8
 
 ### Patch Changes