+++ backend/packages/integrations/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @cat-factory/integrations
 
+## 0.172.16
+
+### Patch Changes
+
+- Updated dependencies [b75fa3c]
+  - @cat-factory/contracts@0.352.0
+  - @cat-factory/kernel@0.345.0
+
+## 0.172.15
+
+### Patch Changes
+
+- Updated dependencies [bba4beb]
+  - @cat-factory/kernel@0.344.0
+
 ## 0.172.14
 
 ### Patch Changes